refactor(runtime): share browser mount update core
